+++ src/main/kotlin/kaiex/exchange/simulator/adapters/HistDataAdapter.kt
@@ -17,13 +17,13 @@ class HistDataAdapter(val symbol: String) : TickAdapter {
     override fun convert(line: String): Trade {
         val fields = line.split(",")
         val timestampStr = fields[0]
-        val bidPrice = fields[1].toFloat()
-        val askPrice = fields[2].toFloat()
+        val bidPrice = fields[1].toBigDecimal()
+        val askPrice = fields[2].toBigDecimal()
         return Trade(
             symbol = symbol,
             side = if (Random.nextBoolean()) OrderSide.BUY else OrderSide.SELL,
-            size = Random.nextFloat() * 1000,
-            price = (bidPrice * 20000f),
+            size = Random.nextFloat().toBigDecimal() * 1000.0.toBigDecimal(),
+            price = (bidPrice * 20000.0.toBigDecimal()),
             createdAt = LocalDateTime.parse(timestampStr, formatter).atZone(ZoneId.systemDefault()).toInstant(),
             receivedAt = Instant.now(),
             liquidation = false,
